..\..\src\ascent-world\Spell.cpp(3601) : warning C4146: unary minus operator applied to unsigned type, result still unsigned .... does it now! amazing!
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3931 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/Spell.cpp
@@ -3598,7 +3598,7 @@ exit:
 	
 	if( u_caster != NULL )
 	{
-		int32 diff = -m_spellInfo->baseLevel;
+		int32 diff = -(int32)m_spellInfo->baseLevel;
 
 		if( m_spellInfo->maxLevel && u_caster->getLevel() > m_spellInfo->maxLevel )
 			diff += m_spellInfo->maxLevel;
